chemical_equipment_visualizer/backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
pandas==2.0.3